+++ zentone/src/main/java/github/nisrulz/zentone/internal/SineWaveGenerator.kt
@@ -4,7 +4,7 @@ import kotlin.math.sin
 
 object SineWaveGenerator : WaveByteArrayGenerator {
 
-    override fun calculateData(index: Int, samplingInterval: Int, amplitude: Int): Byte {
+    override fun calculateData(index: Int, samplingInterval: Float, amplitude: Int): Byte {
         val angle = (Math.PI * index) / samplingInterval
         return (amplitude * sin(angle) * Byte.MAX_VALUE).toInt().toByte()
     }
